backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# ML and Image Processing
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
pillow==10.1.0
opencv-python==4.8.1.78
albumentations==1.3.1
numpy==1.24.3

# Storage
boto3==1.29.7
cloudinary==1.36.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.1

# Utilities
httpx==0.25.2
aiofiles==23.2.1
python-slugify==8.0.1
email-validator==2.1.0

# Monitoring and Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0